Use pick.json and remove get_team_info1 kludge
Store the picks as json so that we don't need to reformat and convert
each time the picks are read.  Also remove the filter around
get_team_info that added games between round 1 and round 2.  If we
need to test this code during the offseason, a more generic solution
should be found.